Merge pull request #56 from lovetriangle2026/feature/course

+++ LoveTriangle-LMS/src/main/java/com/module1/crud/course/controller/CourseController.java
@@ -18,7 +18,7 @@ public class CourseController {
         return service.findAllCourses();
     }
 
-    public List<CourseDTO> findMyCourses(Long userId) throws SQLException {
+    public List<CourseDTO> findMyCourses(int userId) throws SQLException {
         return service.findMyCourses(userId);
     }
